% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aqueduct.R
\name{aqueduct}
\alias{aqueduct}
\title{Run an aqueduct Workflow}
\usage{
aqueduct(x, y)
}
\value{
A dataframe of all of the workflow nodes and if they have been
activated
}
\description{
This function runs the workflow provided in the arguments, ignoring
components of the workflow (nodes) whose parent nodes have not changed
}
