% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aqueduct_initial.R
\name{aqueduct_initial}
\alias{aqueduct_initial}
\title{Initial setup of an aqueduct workflow}
\usage{
aqueduct_initial(basepath, ..., projname)
}
\arguments{
\item{basepath}{Character string; directory in which you want the aqueduct files to be
located, provided as an exact formula (see entry below).
If \code{update=TRUE} then this parameter need not be provided.}

\item{...}{Filepath formulas, separated by commas. There are two kinds of formulas.
An exact formula should be given as a string. These work best for
establishing the basepaths from the root of your system from which
all other paths are derived. For example,
\deqn{basepath = "C:/Users/example_user"}
The formulas may also be given as relative paths, where the first level
is derived from previously defined filepaths. For example,
\deqn{datapath ~ basepath/data/raw}
which would correspond to "C:/Users/example_user/data/raw"}

\item{projname}{Character string; \code{"aqueduct"} by default.
The informative name you want to give to the aqueduct workflow
project. The choice of this parameter has little practical implications
other than to make statements printed from aqueduct() more informative}
}
\description{
Helper function for aqueduct_setup()
}
