% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aqueduct_setup.R
\name{aqueduct_setup}
\alias{aqueduct_setup}
\title{Run an aqueduct Workflow}
\usage{
aqueduct_setup(...)
}
\arguments{
\item{outdir(outdata)~nodename(dir1(indata1)+dir2(indata2)+...)}{A formula instructing which datasets (e.g. \code{indata1})
a workflow node (\code{nodename}) accepts as input}
}
\value{
A dataframe of all of the workflow nodes and if they have been
activated
}
\description{
This function runs the workflow provided in the arguments, ignoring
components of the workflow (nodes) whose parent nodes have not changed
}
